+++ include/offset/distribution/offset_distribution.h
@@ -6,14 +6,14 @@
 class OffsetDistribution {
 
 protected:
-    const int limit;
-    const int block_size;
+    const size_t limit;
+    const size_t block_size;
 
 public:
     OffsetDistribution() = delete;
-    explicit OffsetDistribution(int limit, int block_size);
+    explicit OffsetDistribution(size_t limit, size_t block_size);
     virtual ~OffsetDistribution() = default;
-    virtual int sample() = 0;
+    virtual size_t sample() = 0;
 };
 
 #endif
