+++ src/offset/distribution/sequential_offset.cpp
@@ -1,9 +1,9 @@
 #include <offset/distribution/sequential_offset.h>
 
-SequentialOffset::SequentialOffset(int limit, int block_size)
+SequentialOffset::SequentialOffset(size_t limit, size_t block_size)
     : OffsetDistribution(limit, block_size), current_offset(0) {}
 
-int SequentialOffset::sample() {
+size_t SequentialOffset::sample() {
     int offset = this->current_offset;
     this->current_offset = (this->current_offset + this->block_size) % this->limit;
     return offset;
